django/requirements.txt
Django
psycopg2
djangorestframework
markdown
django-filter
python-social-auth
ipython
